Refactor/router api

+++ app/routers/calendar.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, status, HTTPException, Depends
 from sqlalchemy import select
-from database.db_connection import SessionFactory, get_db
+from database.db_connection import get_db
 from models.user import User
 from models.calendar import Calendar
-from auth.password import hash_password, verify_password
-from schema.calendar.calendar_request import CalendarCreateRequest, CalendarUpdateRequest
+from schema.calendar.calendar_request import CalendarCreateRequest
 from schema.calendar.calendar_response import CalendarResponse
 from sqlalchemy.orm import Session
+from exceptions.handler import NotFoundException, ConflictException
 
 router = APIRouter(tags=["Calendar"])
 
@@ -18,27 +18,23 @@ router = APIRouter(tags=["Calendar"])
     summary = "전체 Calendar 조회"
 )
 def get_calendars_handler(user_id: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendars = session.query(Calendar).filter(
-        Calendar.user_id == user_id
-    ).all()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendars = session.execute(
+        select(Calendar).where(
+            Calendar.user_id == user_id
+        )
+    ).scalars().all()
 
     if not existing_calendars:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 달력이 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
     return existing_calendars
 
@@ -50,28 +46,24 @@ def get_calendars_handler(user_id: int, session : Session = Depends(get_db)):
     summary = "단일 Calendar 조회"
 )
 def get_calendar_handler(user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
     return existing_calendar
 
@@ -83,29 +75,26 @@ def get_calendar_handler(user_id: int, year: int, session : Session = Depends(ge
     summary = "단일 Calendar 생성"
 )
 def create_calendar_handler(body: CalendarCreateRequest, user_id: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #동일한 연도의 달력이 존재하는지 중복 검사 (409 CONFLICT)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.user_id == user_id,
-        Calendar.year == body.year
-    ).first()
+    #동일한 year 의 Calendar 가 존재하는지 중복 검사 (409 CONFLICT)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.user_id == user_id,
+            Calendar.year == body.year
+        )
+    ).scalar_one_or_none()
 
     if existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="이미 존재하는 달력입니다."
-        )
+        raise ConflictException("이미 존재하는 Calendar 입니다.")
 
+    #Calendar 추가
     calendar = Calendar(
         user_id = user_id,
         year = body.year
@@ -124,30 +113,26 @@ def create_calendar_handler(body: CalendarCreateRequest, user_id: int, session :
     summary = "단일 Calendar 삭제"
 )
 def delete_calendar_handler(year: int, user_id: int, session : Session = Depends(get_db)):
-
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
+    #확인된 Calendar 삭제
     session.delete(existing_calendar)
     session.commit()
 
@@ -160,23 +145,28 @@ def delete_calendar_handler(year: int, user_id: int, session : Session = Depends
     summary = "전체 Calendar 삭제"
 )
 def delete_calendars_handler(user_id: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
+        raise NotFoundException("존재하지 않는 User 입니다.")
+
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendars = session.execute(
+        select(Calendar).where(
+            Calendar.user_id == user_id
         )
+    ).scalars().all()
 
-    stmt = select(Calendar).where(Calendar.user_id == user_id)
-    calendars = session.execute(stmt).scalars().all()
+    if not existing_calendars:
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    for calendar in calendars:
+    #확인된 Calendar 전부 삭제
+    for calendar in existing_calendars:
         session.delete(calendar)
     
     session.commit()
 
-    return
+    return None
