+++ app/routers/daily.py
@@ -1,12 +1,13 @@
 from fastapi import APIRouter, status, HTTPException, Depends
 from sqlalchemy import select
-from database.db_connection import SessionFactory, get_db
+from database.db_connection import get_db
 from models.user import User
 from models.calendar import Calendar
 from models.daily import Daily
 from sqlalchemy.orm import Session
 from schema.daily.daily_request import DailyCreateRequest, DailyUpdateRequest
 from schema.daily.daily_response import DailyResponse
+from exceptions.handler import NotFoundException, ConflictException
 
 router = APIRouter(tags=["Daily"])
 
@@ -18,32 +19,31 @@ router = APIRouter(tags=["Daily"])
     summary = "전체 Daily 조회"
 )
 def get_dailies_handler(user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 사용자가 없습니다."
-        )
-    
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.user_id == user_id,
-        Calendar.year == year
-    ).first()
+        raise NotFoundException("존재하지 않는 User 입니다.")
+
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+            select(Calendar).where(
+                Calendar.user_id == user_id,
+                Calendar.year == year
+            )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 달력이 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    existing_dailies = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id
-    ).all()
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
+    existing_dailies = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id
+        )
+    ).scalars().all()
 
     return existing_dailies
 
@@ -55,41 +55,36 @@ def get_dailies_handler(user_id: int, year: int, session : Session = Depends(get
     summary = "단일 Daily 조회"
 )
 def get_daily_handler(user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 사용자가 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.user_id == user_id,
-        Calendar.year == year
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.user_id == user_id,
+            Calendar.year == year
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 달력이 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
-    existing_daily = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id,
-        Daily.month == month,
-        Daily.day == day
-    ).first()
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
+    existing_daily = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id,
+            Daily.month == month,
+            Daily.day == day
+        )
+    ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 날짜가 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
     return existing_daily
 
@@ -101,42 +96,38 @@ def get_daily_handler(user_id: int, year: int, month: int, day: int, session : S
     summary = "단일 Daily 생성"
 )
 def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #동일한 날짜가 존재하는지 중복 검사 (409 CONFLICT)
-    existing_daily = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id,
-        Daily.month == body.month,
-        Daily.day == body.day
-    ).first()
+    #동일한 Daily 가 존재하는지 중복 검사 (409 CONFLICT)
+    existing_daily = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id,
+            Daily.month == body.month,
+            Daily.day == body.day
+        )
+    ).scalar_one_or_none()
 
     if existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="이미 존재하는 날짜입니다."
-        )
+        raise ConflictException("이미 존재하는 Daily 입니다.")
 
+    #Daily 추가
     daily = Daily(
         calendar_id = existing_calendar.calendar_id,
         month = body.month,
@@ -156,43 +147,38 @@ def create_daily_handler(body: DailyCreateRequest, user_id: int, year: int, sess
     summary = "단일 Daily 삭제"
 )
 def delete_daily_handler(user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
-    existing_daily = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id,
-        Daily.month == month,
-        Daily.day == day
-    ).first()
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
+    existing_daily = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id,
+            Daily.month == month,
+            Daily.day == day
+        )
+    ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하는 날짜가 없습니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
+    #Daily 삭제
     session.delete(existing_daily)
     session.commit()
 
@@ -204,35 +190,37 @@ def delete_daily_handler(user_id: int, year: int, month: int, day: int, session 
     status_code=status.HTTP_204_NO_CONTENT,
     summary = "전체 Daily 삭제"
 )
-def delete_daily_handler(user_id: int, year: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+def delete_dailies_handler(user_id: int, year: int, session : Session = Depends(get_db)):
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
 
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
+
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
+    existing_dailies = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id
         )
+    ).scalars().all()
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
-    existing_dailies = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id
-    ).all()
+    if not existing_dailies:
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
+    #확인된 Dailies 삭제
     for existing_daily in existing_dailies:
         session.delete(existing_daily)
 
@@ -245,58 +233,53 @@ def delete_daily_handler(user_id: int, year: int, session : Session = Depends(ge
     "/users/{user_id}/calendars/{year}/dailies/{month}/{day}",
     response_model = DailyResponse,
     status_code = status.HTTP_200_OK,
-    summary = "단일 Daily 생성"
+    summary = "단일 Daily 수정"
 )
 def update_daily_handler(body: DailyUpdateRequest, user_id: int, year: int, month: int, day: int, session : Session = Depends(get_db)):
-    #존재하는 사용자인지 검사 (404 NOT FOUND)
+    #존재하는 User 인지 검사 (404 NOT FOUND)
     existing_user = session.execute(
         select(User).where(User.user_id == user_id)
     ).scalar_one_or_none()
 
     if not existing_user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 사용자입니다."
-        )
+        raise NotFoundException("존재하지 않는 User 입니다.")
     
-    #존재하는 달력인지 검사 (404 NOT FOUND)
-    existing_calendar = session.query(Calendar).filter(
-        Calendar.year == year,
-        Calendar.user_id == user_id
-    ).first()
+    #존재하는 Calendar 인지 검사 (404 NOT FOUND)
+    existing_calendar = session.execute(
+        select(Calendar).where(
+            Calendar.year == year,
+            Calendar.user_id == user_id
+        )
+    ).scalar_one_or_none()
 
     if not existing_calendar:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 달력입니다."
-        )
+        raise NotFoundException("존재하지 않는 Calendar 입니다.")
 
-    #존재하는 날짜인지 검사 (404 NOT FOUND)
-    existing_daily = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id,
-        Daily.month == month,
-        Daily.day == day
-    ).first()
+    #존재하는 Daily 인지 검사 (404 NOT FOUND)
+    existing_daily = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id,
+            Daily.month == month,
+            Daily.day == day
+        )
+    ).scalar_one_or_none()
 
     if not existing_daily:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="존재하지 않는 날짜입니다."
-        )
+        raise NotFoundException("존재하지 않는 Daily 입니다.")
 
-    #바꿀 날짜가 이미 존재하는 날짜인지 검사 (409 CONFLICT)
-    duplicate_daily = session.query(Daily).filter(
-        Daily.calendar_id == existing_calendar.calendar_id,
-        Daily.month == body.month,
-        Daily.day == body.day
-    ).first()
+    #바꿀 Daily 가 이미 존재하는 Daily 인지 검사 (409 CONFLICT)
+    duplicate_daily = session.execute(
+        select(Daily).where(
+            Daily.calendar_id == existing_calendar.calendar_id,
+            Daily.month == body.month,
+            Daily.day == body.day
+        )
+    ).scalar_one_or_none()
 
     if duplicate_daily:
-        raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="이미 존재하는 날짜입니다."
-        )
+        raise ConflictException("이미 존재하는 Daily 입니다.")
 
+    #Daily 수정
     existing_daily.month = body.month
     existing_daily.day = body.day
